Set INTERNAL variables corresponding to Config file
This should allow the project to be added via `add_subdirectory` and
`FetchContent` and then used in the same way (with access to the same
variables) as if it were added via `find_package`.

These variables are primarily used to make decisions about the data
files and how to find them.